Extract core tool models
