+++ docs/DECISIONS.md
@@ -8,7 +8,8 @@
 - TrustedRouter API-key resolution lives in `TrustedRouterAPIKeyResolver`. Developer override precedence, whitespace trimming, session-store fallback, and the actionable missing-key error should stay there so action and safety clients cannot drift.
 - TrustedRouter action and safety transports are separate files. `TrustedRouterLLMClient` owns action streaming; `TrustedRouterSafetyModelClient` owns Auto-review model calls. They share API-key resolution through `TrustedRouterAPIKeyResolver` and JSON response payloads through `TrustedRouterChatParameters` so neither transport depends on the other.
 - TrustedRouter model defaults and catalog normalization live in focused core files, not the general `Models.swift` domain bucket. `TrustedRouterDefaults.swift` owns provider IDs, aliases, bundled fallback rows, Nike 1.0/Prometheus 1.0 branding, and sort/category policy; `ModelInfo.swift` owns the catalog value records.
-- App configuration is a focused core boundary. `AppConfig.swift` owns persisted model/mode/base-URL settings, OAuth/developer-override compatibility, signed-in TrustedRouter account metadata, and favorite-model normalization; `Models.swift` stays focused on general thread/project/tool records.
+- App configuration is a focused core boundary. `AppConfig.swift` owns persisted model/mode/base-URL settings, OAuth/developer-override compatibility, signed-in TrustedRouter account metadata, and favorite-model normalization; `Models.swift` stays focused on general thread/project records.
+- Core tool records are a focused schema boundary. `ToolModels.swift` owns `ToolDefinition`, `ToolCall`, `ToolResult`, built-in core tool definitions, tool-call transcript redaction, and tool-specific browser/memory output compatibility; general chat/thread/project models should not own tool payload compatibility rules.
 - Tool-router feature families should delegate before they sprawl. `ToolRouter` owns the common shell/file/patch entry point and composes git definitions from `GitToolCallDispatcher`; local git, GitHub PR, and worktree tool-call argument mapping lives in that dispatcher. This keeps the shared router small while preserving one public `ToolRouter.execute` API for the agent runtime.
 - Shell request policy is a tool-family concern. `ShellToolCallDispatcher` owns `host.shell.run` definitions, cwd containment, timeout bounds, environment override validation, and `ShellExecutionRequest` construction; `ToolRouter` only delegates shell calls before handling file and patch primitives.
